refactor: isolate timer offboarding
